chore: remove test mode and refactor
